+++ src/django_resume/urls.py
@@ -1,6 +1,15 @@
-from django.urls import path
+from django.urls import path, reverse
+from django.views.generic import RedirectView
 
 from . import views
+
+
+class CvRedirectView(RedirectView):
+    permanent = True
+
+    def get_redirect_url(self, *args, **kwargs):
+        slug = kwargs["slug"]
+        return reverse("django_resume:cv", kwargs={"slug": slug})
 
 
 app_name = "django_resume"
@@ -8,5 +17,6 @@ urlpatterns = [
     path("", views.resume_list, name="list"),
     path("<slug:slug>/delete/", views.resume_delete, name="delete"),
     path("<slug:slug>/", views.resume_detail, name="detail"),
-    path("cv/<slug:slug>/", views.resume_cv, name="cv"),
+    path("<slug:slug>/cv/", views.resume_cv, name="cv"),
+    path("cv/<slug:slug>/", CvRedirectView.as_view(), name="cv-redirect"),
 ]