+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added a resume detail page used as a cover letter
 - Added add resume button to the main page
 - Added a base template for the resume pages
+- Do not require token for CV when user is logged in
 
 0.1.2 - 2024-10-11
 ==================